requirements.txt
absl-py
coverage
flax
grpcio
jax
jaxlib
numpy
portpicker
prometheus-client
pytest
seqio
tiktoken
blobfile
parameterized
shortuuid
fastapi
uvicorn
nltk
evaluate
transformers
# For profiling
tensorboard-plugin-profile